% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreeFunctions.R
\name{MC.Beta}
\alias{MC.Beta}
\title{Monte Carlo estimate of the Beta(a,b) cdf}
\usage{
MC.Beta(x, a, b)
}
\arguments{
\item{x}{the necessary value you need to assign before.}

\item{a}{the first parameter of beta distribution}

\item{b}{the second parameter of beta distribution}
}
\value{
a estimate value of the CDF of the beta(a,b)
}
\description{
Write a function to compute a Monte Carlo estimate of the Beta(a,b) cdf
}
\details{
the first function
}
\examples{
\dontrun{
t <- MC.Beta(0.2,3,3)
t
}
}
