% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreeFunctions.R
\name{my.chisq.test}
\alias{my.chisq.test}
\title{a faster version of chisq.test.}
\usage{
my.chisq.test(x, y)
}
\arguments{
\item{x}{one of the interger vectors of inputs}

\item{y}{one of the interger vectors of inputs}
}
\value{
a estimate value of the CDF of the beta(a,b)
}
\description{
Make a faster version of chisq.test that only computes the chi-square test statistic when the input is two numeric vectors with no missing values.
}
\details{
the second function
}
\examples{
\dontrun{
t <- my.chisq.test(C(1,2,3)),c(4,3,5))
t
}
}
