% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreeFunctions.R
\name{my.pcauchy}
\alias{my.pcauchy}
\title{A function to compute the CDF of the Cauchy distribution}
\usage{
my.pcauchy(x, eta, theta)
}
\arguments{
\item{x}{a value that you need to specified}

\item{eta}{the known parameter with the specific value}

\item{theta}{the known positive parameter in the density}
}
\value{
estimateF= my.pcauchy(x, eta, theta)
}
\description{
Write a function to compute the cdf of the Cauchy distribution, which has the specific density.
}
\details{
the third function
}
\examples{
\dontrun{
estimateF = my.pcauchy(4, 0, 2)
estimateF
}
}
